src/vid2sim_rl/requirements.txt
transformers==4.43.3
wandb==0.17.5
gdown==5.2.0
gymnasium==0.29.1
loguru==0.7.2
numpy==1.23.5
hydra-core==1.3.2
omegaconf==2.3.0
opencv_python==4.10.0.84
pandas==2.2.2
Pillow==10.4.0
setuptools==70.1.1
stable_baselines3==2.3.2
mlagents_envs==1.1.0
